src/tox_requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-environ==0.4.5
django-filter==2.4.0
mysqlclient==2.0.1
pillow==7.1.0
pytest-django==3.10.0
flake8==3.7.9
